html/canvas/element/shadows/2d.shadow.canvasfilter.object.exceptions.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.shadow.canvasfilter.object.exceptions</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.canvasfilter.object.exceptions</h1>
<p class="desc">Test exceptions on CanvasFilter() object</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test exceptions on CanvasFilter() object");
_addTest(function(canvas, ctx) {

assert_throws_js(TypeError, function() { ctx.filter = new CanvasFilter({blur: null}); });
assert_throws_js(TypeError, function() { ctx.filter = new CanvasFilter(null); });
assert_throws_js(TypeError, function() { ctx.filter = new CanvasFilter({blur: undefined}); });
assert_throws_js(TypeError, function() { ctx.filter = new CanvasFilter({foo: {}}); });
assert_throws_js(TypeError, function() { ctx.filter = new CanvasFilter({blur: {}}); });


});
</script>
